+++ controllers/transactions/index.js
@@ -1,7 +1,7 @@
 const addTransaction = require('./add-transactions');
 const removeTransaction = require('./remove-transactionsById');
 const updateTransaction = require('./update-transactionById');
-const getTransactionsByMonth = require('./get-transactionsByMonth ');
+const getTransactionsByMonth = require('./get-transactionsByMonth');
 
 module.exports = {
   addTransaction,
